refactor(zsh): gate install side effects behind DOTFILES_BOOTSTRAP

+++ zsh/go.sh
@@ -1,3 +1,8 @@
 # Import Go environment variables
-. ~/.asdf/plugins/golang/set-env.zsh
-go install golang.org/x/tools/gopls@latest
+if [ -f ~/.asdf/plugins/golang/set-env.zsh ]; then
+  . ~/.asdf/plugins/golang/set-env.zsh
+fi
+
+if exists go && [ "${DOTFILES_BOOTSTRAP:-false}" = "true" ]; then
+  go install golang.org/x/tools/gopls@latest
+fi
